% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generer_rapport.R
\name{generer_rapport}
\alias{generer_rapport}
\title{Generer un rapport personnalise}
\usage{
generer_rapport(commune, departement, output)
}
\arguments{
\item{commune}{La commune pour laquelle generer le rapport (code INSEE).}

\item{departement}{Le departement concerne (code numerique).}

\item{output}{Le nom du fichier de sortie (sans chemin).}
}
\value{
Le chemin du fichier genere.
}
\description{
Cette fonction genere un rapport Quarto au format PDF incluant :
\itemize{
\item Une visualisation des elus par code professionnel pour la commune et le departement.
\item Un resume des informations pour la commune et le departement.
}
}
