% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_commune.R
\name{summary.commune}
\alias{summary.commune}
\title{Résumé pour les objets de classe commune}
\usage{
\method{summary}{commune}(obj, ...)
}
\arguments{
\item{obj}{Un objet de classe \code{commune}.}

\item{...}{Arguments supplémentaires (non utilisés).}
}
\value{
Un résumé des informations de la commune sous forme de liste.
}
\description{
Cette fonction génère un résumé pour les objets de classe \code{commune}.
Elle affiche le nom de la commune et le nombre total d'élus.
}
