% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_distribution_age.R
\name{calcul_distribution_age}
\alias{calcul_distribution_age}
\title{Calculer la distribution des âges des élus}
\usage{
calcul_distribution_age(data)
}
\arguments{
\item{data}{Un dataframe contenant au moins la colonne suivante :
\itemize{
\item \code{Date.de.naissance} (Date) : La date de naissance des élus, qui doit être déjà convertie en format \code{Date}.
}}
}
\value{
Un vecteur numérique représentant les âges des élus en années.
}
\description{
Cette fonction calcule la distribution des âges à partir d'une colonne \code{Date.de.naissance} en format \strong{Date}.
}
\details{
\itemize{
\item La fonction vérifie la présence de la colonne \code{Date.de.naissance}.
\item Elle vérifie que la colonne est bien de type \strong{Date}, sinon elle renvoie une erreur.
\item Elle calcule l'âge en années en utilisant la différence entre la date du jour (\code{Sys.Date()}) et la \code{Date.de.naissance}.
}
}
\examples{
\dontrun{
library(dplyr)
library(lubridate)

# Exemple de dataframe avec la colonne en format Date
data <- data.frame(
  Date.de.naissance = as.Date(c("1950-05-12", "1965-09-23", "1948-11-07"))
)

# Calculer la distribution des âges
calcul_distribution_age(data)
}


}
