% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compter_nombre_d_adjoints.R
\name{compter_nombre_d_elus}
\alias{compter_nombre_d_elus}
\title{Compter le nombre d'élus uniques}
\usage{
compter_nombre_d_elus(data)
}
\arguments{
\item{data}{Un dataframe contenant au moins les colonnes \code{Nom.de.l.élu}, \code{Prénom.de.l.élu} et \code{Date.de.naissance}.}
}
\value{
Un entier représentant le nombre total d'élus dans le dataframe.
}
\description{
Cette fonction compte le nombre total d'élus présents dans un dataframe contenant leurs informations.
}
\details{
\itemize{
\item La fonction vérifie la présence des colonnes obligatoires dans \code{data}.
\item Elle compte le nombre total de lignes du dataframe, supposant que chaque ligne représente un élu unique.
}
}
\examples{
\dontrun{
# Exemple de dataframe
data <- data.frame(
  Nom.de.l.élu = c("Durand", "Martin", "Dupont", "Durand"),
  Prénom.de.l.élu = c("Jean", "Sophie", "Paul", "Jean"),
  Date.de.naissance = c("12/05/1950", "23/09/1965", "07/11/1948", "12/05/1950")
)

# Compter le nombre total d'élus
compter_nombre_d_elus(data)
}


}
